Proyecto/notifications_consumer/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.5
django-environ==0.12.0
flake8==7.1.1
flake8-black==0.3.6
idna==3.10
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
requests==2.32.3
resend==2.6.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
